requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
editor==1.6.6
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
importlib_metadata==8.4.0
inquirer==3.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
litellm==1.44.7
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.9.1
numpy==2.1.0
open-interpreter==0.3.8
openai==1.42.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.9.0
pyshark==0.6
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
runs==1.2.2
scapy==2.5.0
Send2Trash==1.8.3
setuptools==74.0.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
termcolor==2.3.0
tiktoken==0.7.0
tokenizers==0.20.0
tokentrim==0.1.13
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
wget==3.2
xmod==1.8.1
yarl==1.9.4
yaspin==3.0.2
zipp==3.20.1
